[NUB-1419-atttempt-2] [NUB-1419] Bump omniauth to latest version
- Add a form with a post request to start the omniauth handshake
- Modify the rhino schema to include the omniauth urls for each provider
- Simplify loading logic for the OmniauthButton component

Because the Omniauth 2 workflow requires a POST request to begin the handshake,
we can't just redirect the user to the previous endpoint.
The backend schema has been modified to also include the endpoints that start the handshake for each provider.
The frontend will then submit a post request for these providers and omniauth will pick it up from there.

[NUB-1419](https://nubinary.atlassian.net/browse/NUB-1419)